+++ packages/stencil-library/components/tnw-text.js
@@ -2,7 +2,7 @@
  * Built with Stencil
  * Copyright (c) Tecchnway.biz.
  */
-import { T as TnwText$1, d as defineCustomElement$1 } from './p-abe34d52.js';
+import { T as TnwText$1, d as defineCustomElement$1 } from './p-6191d268.js';
 
 const TnwText = TnwText$1;
 const defineCustomElement = defineCustomElement$1;